PCHR-2004: Move absence tab backstopjs files around
