+++ uk.co.compucorp.civicrm.hrcore/backstop_data/casper_scripts/contact-summary/absence/tab-calendar-all-months.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var page = require('../../../page-objects/tabs/absence/calendar');
+var page = require('../../../page-objects/absence-tabs/calendar');
 
 module.exports = function (casper) {
   page.init(casper).showAllMonths();
