+++ uk.co.compucorp.civicrm.hrcore/backstop_data/casper_scripts/contact-summary/absence/tab-calendar.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var page = require('../../../page-objects/tabs/absence/workpatterns');
+var page = require('../../../page-objects/absence-tabs/calendar');
 
 module.exports = function (casper) {
   page.init(casper);
