+++ uk.co.compucorp.civicrm.hrcore/backstop_data/casper_scripts/contact-summary/absence/tab-entitlements.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var page = require('../../../page-objects/tabs/absence/entitlements');
+var page = require('../../../page-objects/absence-tabs/entitlements');
 
 module.exports = function (casper) {
   page.init(casper).waitForReady();
