+++ uk.co.compucorp.civicrm.hrcore/backstop_data/casper_scripts/contact-summary/absence/tab-work-patterns.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var page = require('../../../page-objects/tabs/absence/calendar');
+var page = require('../../../page-objects/absence-tabs/work-patterns');
 
 module.exports = function (casper) {
   page.init(casper);
