+++ uk.co.compucorp.civicrm.hrcore/backstop_data/page-objects/absence-tabs/calendar.js
@@ -1,4 +1,4 @@
-var sspMyLeaveCalendar = require('./../../ssp-leave-absences-my-leave-calendar');
+var sspMyLeaveCalendar = require('../ssp-leave-absences-my-leave-calendar');
 
 module.exports = (function () {
   return sspMyLeaveCalendar.extend({
