+++ uk.co.compucorp.civicrm.hrcore/backstop_data/page-objects/absence-tabs/entitlements.js
@@ -1,4 +1,4 @@
-var page = require('../../contact-summary');
+var page = require('../page');
 
 module.exports = (function () {
   return page.extend({
