+++ uk.co.compucorp.civicrm.hrcore/backstop_data/page-objects/absence-tabs/work-patterns.js
@@ -1,4 +1,4 @@
-var page = require('./../../page');
+var page = require('../page');
 
 module.exports = (function () {
   return page.extend({
