--- uk.co.compucorp.civicrm.hrcore/backstop_data/scenarios/la-admin-leave-calendar.json
@@ -1,14 +1,0 @@
-{
-  "scenarios": [
-    {
-      "label": "Current Month: Admin Leave Calendar",
-      "url": "civicrm/contact/view?reset=1&cid=2&selectedChild=absence#/",
-      "onReadyScript": "contact-summary/absence/tab-calendar_current-month-visible"
-    },
-    {
-      "label": "All Months: Admin Leave Calendar",
-      "url": "civicrm/contact/view?reset=1&cid=2&selectedChild=absence#/",
-      "onReadyScript": "contact-summary/absence/tab-calendar_all-months-visible"
-    }
-  ]
-}
