--- uk.co.compucorp.civicrm.hrcore/backstop_data/scenarios/la-admin-workpatterns.json
@@ -1,14 +1,0 @@
-{
-  "scenarios": [
-    {
-      "label": "Contact Absence / Work Patterns",
-      "url": "civicrm/contact/view?reset=1&cid=2&selectedChild=absence#/",
-      "onReadyScript": "contact-summary/absence/tab-workpatterns"
-    },
-    {
-      "label": "Contact Absence / Custom Work Patterns Modal",
-      "url": "civicrm/contact/view?reset=1&cid=2&selectedChild=absence#/",
-      "onReadyScript": "contact-summary/absence/tab-workpatterns-modal"
-    }
-  ]
-}
